backend/requirements.txt
fastapi
uvicorn[standard]
numpy
pydantic
motor
pymongo
python-dotenv
websockets
python-multipart
hmmlearn
qiskit
qiskit-aer
